refactor: split the files this branch pushed over the 500-LoC cap
CI's file-size-cap job caps .rs files at 500 lines. Five files crossed
it as this branch grew; split each by exact piecewise move (no behavior
change), per the repo's sub-module idiom rather than new allow-list
entries:

  ops/gemm_dense.rs        609 -> 337  (+ gemm_fp4.rs, gemm_fp8_prefill.rs)
  ops/moe_grouped_a.rs     518 -> 479  (+ moe_grouped_fp4.rs)
  nemotron_moe.rs          677 -> 457  (+ prefill_weights.rs, ptr_tables.rs)
  nemotron_mamba2/
    trait_impl.rs          657 -> 211  (+ prefill.rs; the trait method now
                                        delegates to prefill_ssm)
  spark-runtime/buffers.rs 512 -> 232  (+ buffers/accessors.rs)

Also fixes the clippy lints this branch had never been checked against
(CI runs clippy --workspace --tests as errors): manual %-modulo where
is_multiple_of applies, a collapsible if in the shared-expert FP8 gate,
two redundant usize casts, and two dangling #[allow] attributes left by
the extraction.

Verified green locally: cargo fmt --check, clippy --workspace --tests,
cargo test --workspace --locked (1523 pass / 0 fail), SPDX headers on
every new file. Re-verified on GPU after the move: 1k TTFT 405-413ms,
decode 33.0-33.3 tok/s, long-prompt coherence probes clean -- identical
to pre-refactor.
