+++ crates/spark-model/src/layers/ops.rs
@@ -33,6 +33,10 @@ pub mod gdn_flashinfer;
 mod gemm_dense;
 #[path = "ops/gemm_dense_int8.rs"]
 mod gemm_dense_int8;
+#[path = "ops/gemm_fp4.rs"]
+mod gemm_fp4;
+#[path = "ops/gemm_fp8_prefill.rs"]
+mod gemm_fp8_prefill;
 #[path = "ops/gemm_quant.rs"]
 mod gemm_quant;
 #[path = "ops/hyper_connection.rs"]
@@ -57,6 +61,8 @@ mod moe_grouped_a;
 mod moe_grouped_a2;
 #[path = "ops/moe_grouped_b.rs"]
 mod moe_grouped_b;
+#[path = "ops/moe_grouped_fp4.rs"]
+mod moe_grouped_fp4;
 #[path = "ops/moe_prefill.rs"]
 mod moe_prefill;
 #[path = "ops/norm.rs"]
@@ -104,6 +110,8 @@ pub use fp8_moe_batch_a::*;
 pub use fp8_moe_batch_b::*;
 pub use gemm_dense::*;
 pub use gemm_dense_int8::*;
+pub use gemm_fp4::*;
+pub use gemm_fp8_prefill::*;
 pub use gemm_quant::*;
 pub use hyper_connection::*;
 pub use kv_cache::*;
@@ -117,6 +125,7 @@ pub use moe_grouped_a::*;
 pub use moe_grouped_a2::*;
 #[allow(unused_imports)]
 pub(crate) use moe_grouped_b::*;
+pub use moe_grouped_fp4::*;
 pub use moe_prefill::*;
 pub use norm::*;
 pub use nvfp4_mmq::*;
