+++ crates/spark-model/src/layers/ops/gemm_dense.rs
@@ -171,104 +171,10 @@ pub fn w4a16_gemm(
         .launch(stream)
 }
 
-/// Quantize a BF16 [M, K] matrix to NVFP4 (single-level, scale2=1.0): packed E2M1
-/// `[M, K/2]` + per-group-16 E4M3 scales `[M, K/16]`. Prepares W4A4 prefill
-/// activations. Grid = M rows (one block/row), block 128 (threads stride groups).
-#[allow(clippy::too_many_arguments)]
-pub fn quantize_bf16_to_nvfp4(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    input: DevicePtr,
-    packed_out: DevicePtr,
-    scale_out: DevicePtr,
-    m: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([m, 1, 1])
-        .block([128, 1, 1])
-        .arg_ptr(input)
-        .arg_ptr(packed_out)
-        .arg_ptr(scale_out)
-        .arg_f32(1.0) // scale2 = 1.0 (single-level; activation range fits E4M3 group scales)
-        .arg_u32(m) // kernel's N param = rows = tokens
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// W4A4 NVFP4 prefill GEMM (native FP4 tensor cores, sm_121a). Activation is
-/// pre-quantized NVFP4 (`a_packed`/`a_scale`, scale2=1.0); weight is the native
-/// NVFP4 `QuantizedWeight`. Output BF16 [M, N]. See kernels/.../w4a4_gemm.cu.
-/// Grid: (ceil(N/128), ceil(M/128), 1)  Block: (256, 1, 1).
-#[allow(clippy::too_many_arguments)]
-pub fn w4a4_gemm(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    a_packed: DevicePtr,
-    a_scale: DevicePtr,
-    weight: &QuantizedWeight,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(n, 128), div_ceil(m, 128), 1])
-        .block([256, 1, 1])
-        .arg_ptr(a_packed)
-        .arg_ptr(a_scale)
-        .arg_ptr(weight.weight)
-        .arg_ptr(weight.weight_scale)
-        .arg_ptr(output)
-        .arg_f32(1.0) // scaleA2 (activation single-level)
-        .arg_f32(weight.weight_scale_2) // scaleB2 (weight per-tensor)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
 /// W4A16 GEMM with N_TILE=128: same kernel signature, wider N tile.
 ///
 /// Grid: (ceil(N/128), ceil(M/64), 1)  Block: (128, 1, 1)
 #[allow(clippy::too_many_arguments)]
-/// `w4a4_gemm_mfast`: same W4A4 GEMM with M on the fast grid axis, so the
-/// M-blocks sharing a B panel are co-resident and B streams from DRAM once.
-pub fn w4a4_gemm_mfast(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    a_packed: DevicePtr,
-    a_scale: DevicePtr,
-    weight: &QuantizedWeight,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(m, 128), div_ceil(n, 128), 1])
-        .block([128, 1, 1])
-        .arg_ptr(a_packed)
-        .arg_ptr(a_scale)
-        .arg_ptr(weight.weight)
-        .arg_ptr(weight.weight_scale)
-        .arg_ptr(output)
-        .arg_f32(1.0) // scaleA2 (activation single-level)
-        .arg_f32(weight.weight_scale_2) // scaleB2 (weight per-tensor)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// W4A16 GEMM with N_TILE=128: same kernel signature, wider N tile.
-///
-/// Grid: (ceil(N/128), ceil(M/64), 1)  Block: (128, 1, 1)
-#[allow(clippy::too_many_arguments)]
-
 pub fn w4a16_gemm_n128(
     gpu: &dyn GpuBackend,
     kernel: KernelHandle,
@@ -426,184 +332,5 @@ pub fn w4a16_gemm_n128_m128_bf16(
         .arg_u32(m)
         .arg_u32(n)
         .arg_u32(k)
-        .launch(stream)
-}
-
-/// Pre-dequanted FP8 GEMM (prefill): C = A @ B_fp8.
-///
-/// A: [M, K] BF16, B_fp8: [N, K] FP8 E4M3 (pre-dequanted from NVFP4), C: [M, N] BF16.
-/// Eliminates runtime NVFP4→FP8 dequant — only LOAD + FP8 MMA per K step.
-///
-/// Grid: (ceil(N/128), ceil(M/64), 1)  Block: (128, 1, 1)
-#[allow(clippy::too_many_arguments)]
-pub fn fp8_gemm_n128(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    input: DevicePtr,
-    b_fp8: DevicePtr,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(n, 128), div_ceil(m, 64), 1])
-        .block([128, 1, 1])
-        .arg_ptr(input)
-        .arg_ptr(b_fp8)
-        .arg_ptr(output)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// Pre-dequant NVFP4 → FP8 E4M3.  One-time conversion at model load.
-///
-/// Reads B_packed[N, K/2] + B_scale[N, K/GROUP_SIZE] + scale2 → B_fp8[N, K].
-///
-/// Grid: (ceil(N*K/2 / 256), 1, 1)  Block: (256, 1, 1)
-#[allow(clippy::too_many_arguments)]
-/// `fp8_gemm_t_mfast`: same GEMM as [`fp8_gemm_n128`] with the CTA grid axes
-/// swapped so M is the fast axis. The M-blocks that share a B panel then run
-/// co-resident and read it from L2 instead of DRAM; see the kernel comment.
-pub fn fp8_gemm_n128_mfast(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    input: DevicePtr,
-    b_fp8: DevicePtr,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(m, 64), div_ceil(n, 128), 1])
-        .block([128, 1, 1])
-        .arg_ptr(input)
-        .arg_ptr(b_fp8)
-        .arg_ptr(output)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// Pre-dequant NVFP4 → FP8 E4M3.  One-time conversion at model load.
-///
-/// Reads B_packed[N, K/2] + B_scale[N, K/GROUP_SIZE] + scale2 → B_fp8[N, K].
-///
-/// Grid: (ceil(N*K/2 / 256), 1, 1)  Block: (256, 1, 1)
-#[allow(clippy::too_many_arguments)]
-
-/// `fp8_gemm_t_m128_mfast`: 128-row M tile (2 chunks/CTA), m on the fast axis.
-/// Halves the B panel passes relative to [`fp8_gemm_n128_mfast`].
-pub fn fp8_gemm_m128_mfast(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    input: DevicePtr,
-    b_fp8: DevicePtr,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(m, 128), div_ceil(n, 128), 1])
-        .block([128, 1, 1])
-        .arg_ptr(input)
-        .arg_ptr(b_fp8)
-        .arg_ptr(output)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// Pre-dequant NVFP4 → FP8 E4M3.  One-time conversion at model load.
-///
-/// Reads B_packed[N, K/2] + B_scale[N, K/GROUP_SIZE] + scale2 → B_fp8[N, K].
-///
-/// Grid: (ceil(N*K/2 / 256), 1, 1)  Block: (256, 1, 1)
-#[allow(clippy::too_many_arguments)]
-
-/// `fp8_fp8_gemm_t_m128_mfast`: FP8 A x FP8 B, 128-row M tile, m on the fast
-/// axis. A must already be E4M3 (see `bf16_to_fp8`); the MMA consumed E4M3
-/// either way, so pre-casting A is numerically identical to the BF16-A kernel.
-pub fn fp8_fp8_gemm_m128_mfast(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    input: DevicePtr,
-    b_fp8: DevicePtr,
-    output: DevicePtr,
-    m: u32,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(m, 128), div_ceil(n, 128), 1])
-        .block([128, 1, 1])
-        .arg_ptr(input)
-        .arg_ptr(b_fp8)
-        .arg_ptr(output)
-        .arg_u32(m)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// Pre-dequant NVFP4 → FP8 E4M3.  One-time conversion at model load.
-///
-/// Reads B_packed[N, K/2] + B_scale[N, K/GROUP_SIZE] + scale2 → B_fp8[N, K].
-///
-/// Grid: (ceil(N*K/2 / 256), 1, 1)  Block: (256, 1, 1)
-#[allow(clippy::too_many_arguments)]
-
-pub fn predequant_nvfp4_to_fp8(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    b_packed: DevicePtr,
-    b_scale: DevicePtr,
-    scale2: f32,
-    b_fp8: DevicePtr,
-    n: u32,
-    k: u32,
-    stream: u64,
-) -> Result<()> {
-    let total = n * k / 2;
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(total, 256), 1, 1])
-        .block([256, 1, 1])
-        .arg_ptr(b_packed)
-        .arg_ptr(b_scale)
-        .arg_f32(scale2)
-        .arg_ptr(b_fp8)
-        .arg_u32(n)
-        .arg_u32(k)
-        .launch(stream)
-}
-
-/// Convert BF16 activations to FP8 E4M3 for FP8×FP8 GEMM.
-///
-/// Grid: (ceil(total_elements/2 / 256), 1, 1)  Block: (256, 1, 1)
-pub fn bf16_to_fp8(
-    gpu: &dyn GpuBackend,
-    kernel: KernelHandle,
-    src: DevicePtr,
-    dst: DevicePtr,
-    total_elements: u32,
-    stream: u64,
-) -> Result<()> {
-    let threads_needed = total_elements / 2;
-    KernelLaunch::new(gpu, kernel)
-        .grid([div_ceil(threads_needed, 256), 1, 1])
-        .block([256, 1, 1])
-        .arg_ptr(src)
-        .arg_ptr(dst)
-        .arg_u32(total_elements)
         .launch(stream)
 }
